Optionally omit generating directives when instrumenting
This is used the python side so that we don't need to remove
the directives manually after instrumentation

+++ src/ASTEdits.h
@@ -19,7 +19,7 @@ public:
   RuleActionEditCollector(
       clang::transformer::RewriteRule Rule,
       std::vector<clang::tooling::Replacement> &Replacements,
-      std::map<std::string, int> &FileToNumberMarkerDecls)
+      std::map<std::string, size_t> &FileToNumberMarkerDecls)
       : Rule{Rule}, Replacements{Replacements},
         FileToNumberMarkerDecls{FileToNumberMarkerDecls} {}
   void
@@ -29,7 +29,7 @@ public:
 private:
   clang::transformer::RewriteRule Rule;
   std::vector<clang::tooling::Replacement> &Replacements;
-  std::map<std::string, int> &FileToNumberMarkerDecls;
+  std::map<std::string, size_t> &FileToNumberMarkerDecls;
 };
 
 } // namespace markers
